--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -3,8 +3,7 @@ module github.com/Aureys96/hw10_program_optimization
 go 1.16
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/json-iterator/go v1.1.12
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )